+++ db/query/hit_list.Totals.sql
@@ -4,8 +4,8 @@ select
 from hit_counts
 {{:no_events join paths using (path_id)}}
 where
-	hit_counts.site_id = :site and hour >= :start and hour <= :end
-	{{:no_events and paths.event = 0}}
-	{{:filter and path_id :in (:filter)}}
+	hit_counts.site_id = :site and hour >= :start and hour <= :end and
+	{{:no_events paths.event = 0 and}}
+	:filter
 group by hour
 order by hour asc