+++ db/query/hit_stats.ListLocations.sql
@@ -3,9 +3,7 @@ with x as (
 		substr(location, 0, 3) as loc,
 		sum(count)      as count
 	from location_stats
-	where
-		site_id = :site and day >= :start and day <= :end
-		{{:filter and path_id :in (:filter)}}
+	where site_id = :site and day >= :start and day <= :end and :filter
 	group by loc
 	order by count desc, loc
 	limit :limit offset :offset