+++ db/query/hit_stats.ListSize.sql
@@ -3,8 +3,7 @@ select
 	sum(count)     as count
 from size_stats
 where
-	site_id = :site and day >= :start and day <= :end
-	{{:filter   and path_id :in (:filter)}}
+	site_id = :site and day >= :start and day <= :end and :filter
 	{{:max_size and width != 0 and width > :min_size and width <= :max_size}}
 	{{:empty    and width = 0}}
 group by width