+++ db/query/hit_stats.ListSystems.sql
@@ -3,9 +3,7 @@ with x as (
 		system_id,
 		sum(count) as count
 	from system_stats
-	where
-		site_id = :site and day >= :start and day <= :end
-		{{:filter and path_id :in (:filter)}}
+	where site_id = :site and day >= :start and day <= :end and :filter
 	group by system_id
 	order by count desc
 )